+++ src/decryption/mod.rs
@@ -13,7 +13,6 @@ pub(crate) mod stream;
 pub use decrypt::decrypt;
 pub use read::{
     consume_all_extensions, read_exact_span, read_file_version, read_kdf_iterations,
-    read_reserved_modulo_byte,
 };
 pub use session::extract_session_data;
 pub use stream::{decrypt_ciphertext_stream, StreamConfig};